PROTOCOL CHANGE! activate 32 bit nodenums/packetids

+++ src/mesh/MeshTypes.h
@@ -6,8 +6,8 @@
 #include "mesh.pb.h"
 #include <Arduino.h>
 
-typedef uint8_t NodeNum;
-typedef uint8_t PacketId; // A packet sequence number
+typedef uint32_t NodeNum;
+typedef uint32_t PacketId; // A packet sequence number
 
 #define NODENUM_BROADCAST (sizeof(NodeNum) == 4 ? UINT32_MAX : UINT8_MAX)
 #define ERRNO_OK 0
